+++ CHANGES.md
@@ -2,8 +2,11 @@
 Google-Chrome Erweiterung
 
 # Version History
+0.10 Display error in Firefox solved
+0.9 Icons added, Display error in Firefox solved
+0.8 Sorting of Bookmarks and Folders implemented
 0.7 Add support for seperator line in Firefox bookmarks
-0.6 Fix some smaller issues 
+0.6 Add support for Firefox
 0.5 Add multi-language support (German / English)